Poprawa uporządkowania projektu, nowe pliki i nazwy
