--- main.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-#include "storage.hpp"
-int main() {
-//    std::cout << "Hello, World!" << std::endl;
-//    return 0;
-}
